Nathan

+++ TAP2.0/src/tap2/pkg0/tap.sql
@@ -19,7 +19,7 @@ PRIMARY KEY (sensor_serial)
 CREATE TABLE temperature(
 temp_id serial,
 time_stamp timestamp,
-degrees_c int,
+degrees_c float,
 sensor_serial text,
 location_id text,
 PRIMARY KEY (temp_id)
@@ -27,7 +27,7 @@ PRIMARY KEY (temp_id)
 
 INSERT INTO location (location_id,full_name,abbreviation)VALUES ('123', 'New York', 'ny');
 INSERT INTO sensor (sensor_serial, abbreviation) VALUES ('a1', 'ny');
-INSERT INTO temperature (temp_id,time_stamp,degrees_c,sensor_serial,location_id) VALUES ('1','1999-01-08 04:05:06',40,'a1','123');
+INSERT INTO temperature (temp_id,time_stamp,degrees_c,sensor_serial,location_id) VALUES ('1','1999-01-08 04:05:06',40.0,'a1','123');
 
 Create Role postgres with login;
 Grant all on location to postgres;
